Add session support (stored in MongoDB, with signed cookies for sessionId)
Messed up the code a bit, since partially transitioning in an express-like structure, with app and middlewares. This will need to be cleaned. I do not use express itself, nor connect, but a custom implementation for now, for learning purposes. I will see how that turns out.
It has been pretty tricky to plug the express-session and cookieParser on my custom App. Again, cleaning will come later.

IMPORTANT NOTE: sensitive config should be placed in src/server/env/config.json
This file is .gitignore'd, and there is a default config provided. But a specific config.json need to be put on the prod server.