feat: change DB data init strategy from container start to admin endpoint
Process to setup DB data on first use is described in README.md.
When MongoDB container starts with an empty volume, the admin user is
injected. However, the data is not added anymore and will need an api
call to be imported.

Reason for the choice:
While is is more convenient to get everything ready at start, the use of
an older version of MongoDB did not allow to easily share the data files.
This solution simplifies the ConfigMap in the Helm chart and keeps
functional data inside the server image only.